+++ workbook/lib/screens/auth/login_page.dart
@@ -11,8 +11,8 @@ import 'package:regexed_validator/regexed_validator.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:workbook/constants.dart';
 import 'package:workbook/screens/dashboard.dart';
-import 'package:workbook/screens/guest_ticket.dart';
-import 'package:workbook/screens/otp_verification.dart';
+import 'package:workbook/screens/queries/guest_ticket.dart';
+import 'package:workbook/screens/auth/otp_verification.dart';
 import 'dart:convert';
 import 'package:workbook/widget/first.dart';
 import 'package:workbook/widget/input_field.dart';
