+++ workbook/lib/screens/auth/registration_page.dart
@@ -4,9 +4,9 @@ import 'package:flutter/material.dart';
 import 'package:modal_progress_hud/modal_progress_hud.dart';
 import 'package:page_transition/page_transition.dart';
 import 'package:workbook/constants.dart';
-import 'package:workbook/screens/driver_form.dart';
-import 'package:workbook/screens/employee_cust_form.dart';
-import 'package:workbook/screens/admin_form.dart';
+import 'package:workbook/screens/auth/driver_form.dart';
+import 'package:workbook/screens/auth/employee_cust_form.dart';
+import 'package:workbook/screens/auth/admin_form.dart';
 import 'package:workbook/widget/signup.dart';
 import 'package:workbook/widget/textNew.dart';
 import 'package:http/http.dart' as http;
