+++ workbook/lib/screens/grade_and_divisions/add_grade.dart
@@ -8,7 +8,7 @@ import 'package:fluttertoast/fluttertoast.dart';
 import 'package:modal_progress_hud/modal_progress_hud.dart';
 import 'package:page_transition/page_transition.dart';
 import 'package:workbook/constants.dart';
-import 'package:workbook/screens/view_divisions.dart';
+import 'package:workbook/screens/grade_and_divisions/view_divisions.dart';
 import 'package:workbook/user.dart';
 import 'package:workbook/widget/drawer.dart';
 import 'package:http/http.dart' as http;
