+++ workbook/lib/screens/tasks/view_tasks.dart
@@ -4,8 +4,8 @@ import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 import 'package:modal_progress_hud/modal_progress_hud.dart';
 import 'package:http/http.dart' as http;
-import '../constants.dart';
-import '../user.dart';
+import '../../constants.dart';
+import '../../user.dart';
 
 class ViewTasks extends StatefulWidget {
   @override
